REFACTOR: split healthcheck package into subpackages
