+++ core/internal/infra/api/api_endpoints.go
@@ -4,7 +4,7 @@ import (
 	accesstransport "core/internal/app/access/transport"
 	evaluationtransport "core/internal/app/evaluation/transport"
 	flagtransport "core/internal/app/flag/transport"
-	"core/internal/app/healthcheck"
+	healthchecktransport "core/internal/app/healthcheck/transport"
 	identitytransport "core/internal/app/identity/transport"
 	projecttransport "core/internal/app/project/transport"
 	segmenttransport "core/internal/app/segment/transport"
@@ -24,7 +24,7 @@ func ApplyRoutes(senv *srvenv.Env, r *gin.Engine) {
 	accesstransport.ApplyRoutes(senv, root)
 	flagtransport.ApplyRoutes(senv, root)
 	evaluationtransport.ApplyRoutes(senv, root)
-	healthcheck.ApplyRoutes(senv, root)
+	healthchecktransport.ApplyRoutes(senv, root)
 	identitytransport.ApplyRoutes(senv, root)
 	projecttransport.ApplyRoutes(senv, root)
 	targetingtransport.ApplyRoutes(senv, root)
